- Changed: GameLivingEvent.EnemyKilled is now fired for all group members that aren't attacking the enemy. - Bugfix: Forgot to remove a - in the spell resists fix from last commit. (Thanks to Supgee for reporting) - Bugfix: Razing a tower when it is unclaimed will no longer increase the tower level by 1. (Thanks to arkara for reporting) - Bugfix: Trueshot should now have a distance effect, and be removed from the player after a shot. (Thanks to zonks01 for reporting) - Added: Group kill mission support, owner = the PlayerGroup instance. - Added: Server Property "load_quests" setting this to False will stop the quests from loading and spamming the console with text. - Changed: When an Admin logs in, it will be displayed as a staff member in the PlayerEnterExit script. - Bugfix: /mob respawn command is no longer a ushort, but a real int. - Added: /gtrange command

+++ GameServerScripts/gmcommands/mob.cs
@@ -459,7 +459,7 @@ namespace DOL.GS.Scripts
 						int interval;
 						try
 						{
-							interval = Convert.ToInt16(args[2]);
+							interval = Convert.ToInt32(args[2]);
 							((GameMob)targetMob).RespawnInterval = interval;
 							targetMob.SaveIntoDatabase();
 							client.Out.SendMessage("Mob respawn interval changed to: " + interval, eChatType.CT_System, eChatLoc.CL_SystemWindow);
